// packages/js-core/src/proto/polycentric.ts
export interface PublicKey {
  /**
   * @generated from protobuf field: uint64 key_type = 1
   */
  keyType: bigint;
  /**
   * @generated from protobuf field: bytes key = 2
   */
  key: Uint8Array;
}
/**
 * @generated from protobuf message polycentric.PrivateKey
 */
export interface PrivateKey {
  /**
   * @generated from protobuf field: uint64 key_type = 1
   */
  keyType: bigint;
  /**
   * @generated from protobuf field: bytes key = 2
   */
  key: Uint8Array;
}
/**
 * @generated from protobuf message polycentric.VectorClock
 */
export interface VectorClock {
  /**
   * @generated from protobuf field: repeated uint64 logical_clocks = 1
   */
  logicalClocks: bigint[];
}
/**
 * @generated from protobuf message polycentric.Index
 */
export interface Index {
  /**
   * @generated from protobuf field: uint64 index_type = 1
   */
  indexType: bigint;
  /**
   * @generated from protobuf field: uint64 logical_clock = 2
   */
  logicalClock: bigint;
}
/**
 * @generated from protobuf message polycentric.Indices
 */
export interface Indices {
  /**
   * @generated from protobuf field: repeated polycentric.Index indices = 1
   */
  indices: Index[];
}
/**
 * @generated from protobuf message polycentric.Process
 */
export interface Process {
  /**
   * @generated from protobuf field: bytes process = 1
   */
  process: Uint8Array;
}
/**
 * @generated from protobuf message polycentric.Digest
 */
export interface Digest {
  /**
   * @generated from protobuf field: uint64 digest_type = 1
   */
  digestType: bigint;
  /**
   * @generated from protobuf field: bytes digest = 2
   */
  digest: Uint8Array;
}
/**
 * @generated from protobuf enum polycentric.Digest.DigestType
 */
export enum Digest_DigestType {
  /**
   * @generated from protobuf enum value: Unknown = 0;
   */
  Unknown = 0,
  /**
   * @generated from protobuf enum value: SHA256 = 1;
   */
  SHA256 = 1,
}
/**
 * @generated from protobuf message polycentric.Pointer
 */
export interface Pointer {
  /**
   * @generated from protobuf field: polycentric.PublicKey system = 1
   */
  system?: PublicKey;
  /**
   * @generated from protobuf field: polycentric.Process process = 2
   */
  process?: Process;
  /**
   * @generated from protobuf field: uint64 logical_clock = 3
   */
  logicalClock: bigint;
  /**
   * @generated from protobuf field: polycentric.Digest event_digest = 4
   */
  eventDigest?: Digest;
}
/**
 * @generated from protobuf message polycentric.LWWElementSet
 */
export interface LWWElementSet {
  /**
   * @generated from protobuf field: polycentric.LWWElementSet.Operation operation = 1
   */
  operation: LWWElementSet_Operation;
  /**
   * @generated from protobuf field: bytes value = 2
   */
  value: Uint8Array;
  /**
   * @generated from protobuf field: uint64 unix_milliseconds = 3
   */
  unixMilliseconds: bigint;
}
/**
 * @generated from protobuf enum polycentric.LWWElementSet.Operation
 */
export enum LWWElementSet_Operation {
  /**
   * @generated from protobuf enum value: ADD = 0;
   */
  ADD = 0,
  /**
   * @generated from protobuf enum value: REMOVE = 1;
   */
  REMOVE = 1,
}
/**
 * @generated from protobuf message polycentric.LWWElement
 */
export interface LWWElement {
  /**
   * @generated from protobuf field: bytes value = 1
   */
  value: Uint8Array;
  /**
   * @generated from protobuf field: uint64 unix_milliseconds = 2
   */
  unixMilliseconds: bigint;
}
/**
 * @generated from protobuf message polycentric.Reference
 */
export interface Reference {
  /**
   * @generated from protobuf field: uint64 reference_type = 1
   */
  referenceType: bigint;
  /**
   * @generated from protobuf field: bytes reference = 2
   */
  reference: Uint8Array;
}
// ...
export interface Events {
  /**
   * @generated from protobuf field: repeated polycentric.SignedEvent events = 1
   */
  events: SignedEvent[];
}
/**
 * @generated from protobuf message polycentric.Post
 */
export interface Post {
  /**
   * @generated from protobuf field: string content = 1
   */
  content: string;
  /**
   * @generated from protobuf field: optional polycentric.ImageManifest image = 2
   */
  image?: ImageManifest;
}
/**
 * @generated from protobuf message polycentric.ImageManifest
 */
export interface ImageManifest {
  /**
   * @generated from protobuf field: string mime = 1
   */
  mime: string;
  /**
   * @generated from protobuf field: uint32 width = 2
   */
  width: number;
  /**
   * @generated from protobuf field: uint32 height = 3
   */
  height: number;
  /**
   * @generated from protobuf field: bytes digest = 4
   */
  digest: Uint8Array;
}
// ...
export interface Event {
  /**
   * @generated from protobuf field: polycentric.PublicKey system = 1
   */
  system?: PublicKey;
  /**
   * @generated from protobuf field: polycentric.Process process = 2
   */
  process?: Process;
  /**
   * @generated from protobuf field: uint64 logical_clock = 3
   */
  logicalClock: bigint;
  /**
   * @generated from protobuf field: polycentric.ContentType content_type = 4
   */
  contentType: ContentType;
  /**
   * @generated from protobuf field: bytes content = 5
   */
  content: Uint8Array;
  /**
   * @generated from protobuf field: polycentric.VectorClock vector_clock = 6
   */
  vectorClock?: VectorClock;
  /**
   * @generated from protobuf field: polycentric.Indices indices = 7
   */
  indices?: Indices;
  /**
   * @generated from protobuf field: optional polycentric.LWWElementSet lww_element_set = 8
   */
  lwwElementSet?: LWWElementSet;
  /**
   * @generated from protobuf field: optional polycentric.LWWElement lww_element = 9
   */
  lwwElement?: LWWElement;
  /**
   * @generated from protobuf field: repeated polycentric.Reference references = 10
   */
  references: Reference[];
  /**
   * @generated from protobuf field: optional uint64 unix_milliseconds = 11
   */
  unixMilliseconds?: bigint;
}
// ...
export interface FeedResult {
  /**
   * @generated from protobuf field: repeated polycentric.SignedEvent events = 1
   */
  events: SignedEvent[];
  /**
   * @generated from protobuf field: bytes cursor = 2
   */
  cursor: Uint8Array;
}
// ...
export interface EventKey {
  /**
   * @generated from protobuf field: uint64 system_key_type = 1
   */
  systemKeyType: bigint;
  /**
   * @generated from protobuf field: bytes system_key = 2
   */
  systemKey: Uint8Array;
  /**
   * @generated from protobuf field: bytes process = 3
   */
  process: Uint8Array;
  /**
   * @generated from protobuf field: uint64 logical_clock = 4
   */
  logicalClock: bigint;
}
// ...
export interface ResultEventsAndRelatedEventsAndCursor {
  /**
   * @generated from protobuf field: polycentric.Events result_events = 1
   */
  resultEvents?: Events;
  /**
   * @generated from protobuf field: polycentric.Events related_events = 2
   */
  relatedEvents?: Events;
  /**
   * @generated from protobuf field: optional bytes cursor = 3
   */
  cursor?: Uint8Array;
}
// ...
export enum ContentType {
  /**
   * @generated from protobuf enum value: UNKNOWN = 0;
   */
  UNKNOWN = 0,
  /**
   * @generated from protobuf enum value: DELETE = 1;
   */
  DELETE = 1,
  /**
   * @generated from protobuf enum value: SYSTEM_PROCESSES = 2;
   */
  SYSTEM_PROCESSES = 2,
  /**
   * @generated from protobuf enum value: POST = 3;
   */
  POST = 3,
  /**
   * @generated from protobuf enum value: FOLLOW = 4;
   */
  FOLLOW = 4,
  /**
   * @generated from protobuf enum value: USERNAME = 5;
   */
  USERNAME = 5,
  /**
   * @generated from protobuf enum value: DESCRIPTION = 6;
   */
  DESCRIPTION = 6,
  /**
   * @generated from protobuf enum value: BLOB_META = 7;
   */
  BLOB_META = 7,
  /**
   * @generated from protobuf enum value: BLOB_SECTION = 8;
   */
  BLOB_SECTION = 8,
  /**
   * @generated from protobuf enum value: AVATAR = 9;
   */
  AVATAR = 9,
  /**
   * @generated from protobuf enum value: SERVER = 10;
   */
  SERVER = 10,
  /**
   * @generated from protobuf enum value: VOUCH = 11;
   */
  VOUCH = 11,
  /**
   * @generated from protobuf enum value: CLAIM = 12;
   */
  CLAIM = 12,
  /**
   * @generated from protobuf enum value: BANNER = 13;
   */
  BANNER = 13,
  /**
   * @generated from protobuf enum value: OPINION = 14;
   */
  OPINION = 14,
  /**
   * @generated from protobuf enum value: STORE = 15;
   */
  STORE = 15,
  /**
   * @generated from protobuf enum value: AUTHORITY = 16;
   */
  AUTHORITY = 16,
  /**
   * @generated from protobuf enum value: JOIN_TOPIC = 17;
   */
  JOIN_TOPIC = 17,
  /**
   * @generated from protobuf enum value: BLOCK = 18;
   */
  BLOCK = 18,
}
// ...
class PublicKey$Type extends MessageType<PublicKey> {
  constructor() {
    super('polycentric.PublicKey', [
      {
        no: 1,
        name: 'key_type',
        kind: 'scalar',
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 2, name: 'key', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
    ]);
  }
  create(value?: PartialMessage<PublicKey>): PublicKey {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.keyType = 0n;
    message.key = new Uint8Array(0);
    if (value !== undefined)
      reflectionMergePartial<PublicKey>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: PublicKey,
  ): PublicKey {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* uint64 key_type */ 1:
          message.keyType = reader.uint64().toBigInt();
          break;
        case /* bytes key */ 2:
          message.key = reader.bytes();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: PublicKey,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* uint64 key_type = 1; */
    if (message.keyType !== 0n)
      writer.tag(1, WireType.Varint).uint64(message.keyType);
    /* bytes key = 2; */
    if (message.key.length)
      writer.tag(2, WireType.LengthDelimited).bytes(message.key);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message polycentric.PublicKey
 */
export const PublicKey = new PublicKey$Type();
// ...
export const VectorClock = new VectorClock$Type();
// ...
export const Index = new Index$Type();
// ...
export const Indices = new Indices$Type();
// ...
export const Process = new Process$Type();
// ...
export const Digest = new Digest$Type();
// ...
export const Pointer = new Pointer$Type();
// ...
export const LWWElementSet = new LWWElementSet$Type();
// @generated message type with reflection information, may provide speed optimized methods
class LWWElement$Type extends MessageType<LWWElement> {
  constructor() {
    super('polycentric.LWWElement', [
      { no: 1, name: 'value', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
      {
        no: 2,
        name: 'unix_milliseconds',
        kind: 'scalar',
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
    ]);
  }
  create(value?: PartialMessage<LWWElement>): LWWElement {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.value = new Uint8Array(0);
    message.unixMilliseconds = 0n;
    if (value !== undefined)
      reflectionMergePartial<LWWElement>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: LWWElement,
  ): LWWElement {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* bytes value */ 1:
          message.value = reader.bytes();
          break;
        case /* uint64 unix_milliseconds */ 2:
          message.unixMilliseconds = reader.uint64().toBigInt();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: LWWElement,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* bytes value = 1; */
    if (message.value.length)
      writer.tag(1, WireType.LengthDelimited).bytes(message.value);
    /* uint64 unix_milliseconds = 2; */
    if (message.unixMilliseconds !== 0n)
      writer.tag(2, WireType.Varint).uint64(message.unixMilliseconds);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message polycentric.LWWElement
 */
export const LWWElement = new LWWElement$Type();
// ...
export const Reference = new Reference$Type();
// ...
export const SignedEvent = new SignedEvent$Type();
// ...
export const Events = new Events$Type();
// ...
export const ImageManifest = new ImageManifest$Type();
// ...
export const Event = new Event$Type();
// ...
export const FeedResult = new FeedResult$Type();
// ...
export const EventKey = new EventKey$Type();
// ...
export const ResultEventsAndRelatedEventsAndCursor =
  new ResultEventsAndRelatedEventsAndCursor$Type();